pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tskr"
dynamic = ["version"]
description = "A clean, developer-friendly task management CLI designed for LLM collaboration"
keywords = ["cli", "task-management", "productivity", "git", "llm", "collaboration", "developer-tools"]
license = "MIT"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Bug Tracking",
    "Topic :: Office/Business :: Scheduling",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "python-dateutil>=2.8.0",
    "pyinstaller>=6.16.0",
    "psutil>=7.1.0",
]

[project.urls]
Homepage = "https://github.com/tskr-dev/tskr"
Repository = "https://github.com/tskr-dev/tskr.git"
Issues = "https://github.com/tskr-dev/tskr/issues"
Documentation = "https://github.com/tskr-dev/tskr/wiki"
Changelog = "https://github.com/tskr-dev/tskr/blob/main/CHANGELOG.md"
"Bug Reports" = "https://github.com/tskr-dev/tskr/issues/new?template=bug_report.md"
"Feature Requests" = "https://github.com/tskr-dev/tskr/issues/new?template=feature_request.md"

[project.scripts]
tskr = "tskr.cli:app"

[tool.hatch.version]
path = "src/tskr/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/tskr"]

[tool.uv]
dev-dependencies = [
    "pytest==7.4.4",
    "pyinstaller>=5.0.0",
    "ruff==0.1.8",
    "mypy==1.8.0",
    "types-python-dateutil>=2.8.0",
    "pre-commit>=4.3.0",
    "pytest-cov>=7.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py39"
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "PIE", "SIM"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
